Fix OpenZeppelin imports and decimal math

+++ contracts/src/FortuneAssetRegistry.sol
@@ -1,7 +1,8 @@
 // SPDX-License-Identifier: MIT
 pragma solidity ^0.8.24;
 
-import {Ownable2Step, Ownable} from "@openzeppelin/contracts/access/Ownable2Step.sol";
+import {Ownable2Step} from "@openzeppelin/contracts/access/Ownable2Step.sol";
+import {Ownable} from "@openzeppelin/contracts/access/Ownable.sol";
 import {IERC20Metadata} from "@openzeppelin/contracts/token/ERC20/extensions/IERC20Metadata.sol";
 import {IFortunePriceOracle} from "./interfaces/IFortunePriceOracle.sol";
 
@@ -82,7 +83,7 @@ contract FortuneAssetRegistry is Ownable2Step {
         require(block.timestamp - updatedAt <= config.maxOracleAge, "STALE_PRICE");
 
         uint8 decimals = IERC20Metadata(asset).decimals();
-        return amount * price / (10 ** decimals);
+        return amount * price / (10 ** uint256(decimals));
     }
 
     function tokenAmountForUsd(address asset, uint256 usd1e18) external view returns (uint256 amount) {
@@ -94,7 +95,7 @@ contract FortuneAssetRegistry is Ownable2Step {
         require(block.timestamp - updatedAt <= config.maxOracleAge, "STALE_PRICE");
 
         uint8 decimals = IERC20Metadata(asset).decimals();
-        return usd1e18 * (10 ** decimals) / price;
+        return usd1e18 * (10 ** uint256(decimals)) / price;
     }
 
     function assetCount() external view returns (uint256) {
